requirements.txt
Django==4.1.1
django-extensions==3.2.1
djangorestframework==3.13.1
geographiclib==1.52
geopy==2.2.0
idna==3.4
numpy==1.23.3
pandas==1.4.4
psycopg2==2.9.3
pyproj==3.4.0
requests==2.28.1